SSOR-56 Refactor: File structure changes, fixed a few tests
