+++ src/app/transaction/components/transaction-view/transaction-queries/transaction-queries.component.html
