+++ src/app/transaction/components/transaction-view/transaction-view.component.ts
